changes to fjp.py fixed & moved the code to determine the inner points such that they don't overlap -- one issue is that we're not properly comparing them against the original edge line & thus will fail under some cases
polygon line originally is o1 - - o2
if we inset that line via a calculated constant from the dihedral
angle we'll end up with 'inset' lines  i1 - - i2 (same length as original)
these inside lines intersect with their neighbours so we can find
the intersecting inset lines x1 --- x2
line x & and line i are colinear but we're currently assuming that the order
would be:
   i1 -- x1 ---- x2 -- i2
though in principle this is not necessarily the case; and is an edge condition
we are currently hitting though we actually avoid dealing with it (maybe luckily?)

we are also not dealing with the case of the opposing side lining up terribly so
this may actually be a real issue :( ... what a bummer. I may need to actually fix
this for my own silly project.

  a_i2 -- a_x2 ------- a_x1  ---- a_i1

  o1 -    - - - - - - - - - - - - o2

  b_i1 ------ b_x1 ------ b_x2 -- b_i2

We are currently offsetting by the maximal offset (i.e. between b_i1 & b_x1 in this
case just looking) but if it's different' between ax_2 & ax_1 then we'd be sad :(

-- this points towards generating a list of inset points & then comparing to opposite
half edges -- what a bummer.  Maybe the whole "let the user come up with a universal
offset & just go from there was a better / easier plan.