Major changes to code in here are to help better determine the "offsets" of an edge as mentioned in the last commit message we needed to worry about offsets that are greater than the original length as well as the offsets provided by the opposite edge -- this was fixed by calculating the necessary offsets for each half edge and then using the maximum offset found between he pairs.
I also found the pyclipper library and used it to add kerf offset to the "squaggles"
outline  as well as generate a window (which I used to often generate by hand).  I
was wondering if I could more easily create the "inset" / insideSplines using this
library -- my current thoughts are you'd have to end up creating clipping lines
and somehow keep track of them as stuff gets clipped -- to see if that edge
disappears say... not obvious to me how this would work exactly but likely
possible.

One thing I learned in the pyclipper library is that "JT_ROUND" is somehow
incredibly resource intensive / was generating 60mb files - switching to
alternative joints was better for sure.

Also modified the Polyhedron class to accept some more parameters to do with
material thickness & border_width for windows.  I'd like to add in a few more
parameters here, but this is starting to actually look like a working script..!

The other nice to have would be a better map showing how pieces fit together
it was a little confusing already, and I can only imagine it getting more so if
you attempt a larger project.